requirements.txt
asgiref==3.8.1
boto3==1.35.20
botocore==1.35.20
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.4.0
dj-database-url==2.2.0
Django==5.1.1
django-cors-headers==4.4.0
django-heroku==0.3.1
django-storages==1.14.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.10
jmespath==1.0.1
mysqlclient==2.2.4
pillow==10.4.0
#psycopg2==2.9.10
PyJWT==2.9.0
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
pytz==2024.2
reportlab==4.2.4
requests==2.32.3
s3transfer==0.10.2
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.3
whitenoise==6.7.0
